Move component to the right place
