+++ app/src/ui/changes/sidebar.tsx
@@ -32,7 +32,6 @@ import { getLargeFilePaths } from '../../lib/large-files'
 import { isConflictedFile, hasUnresolvedConflicts } from '../../lib/status'
 import { enablePullWithRebase } from '../../lib/feature-flag'
 import { IStashEntry } from '../../lib/git/stash'
-import { Octicon, OcticonSymbol } from '../octicons'
 
 /**
  * The timeout for the animation of the enter/leave animation for Undo.
@@ -408,24 +407,10 @@ export class ChangesSidebar extends React.Component<IChangesSidebarProps, {}> {
           onOpenInExternalEditor={this.props.onOpenInExternalEditor}
           onChangesListScrolled={this.props.onChangesListScrolled}
           changesListScrollTop={this.props.changesListScrollTop}
+          stashEntry={this.props.stashEntry}
         />
-        {this.renderStashedChanges()}
         {this.renderUndoCommit(rebaseConflictState)}
       </div>
-    )
-  }
-
-  private renderStashedChanges() {
-    if (this.props.stashEntry === null) {
-      return null
-    }
-
-    return (
-      <ul className="brutalism">
-        <li>
-          Stashed Changes <Octicon symbol={OcticonSymbol.arrowRight} />
-        </li>
-      </ul>
     )
   }
 }
